+++ src/components/20-molecules/datepicker/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 7.0.5
+
+- Fix wrong date formatting for locale `it-CH` (#1740)
+- Now all month identifiers of all languages are capitalized at IE browser.
+
 ## 7.0.4
 
 - Remove min-width from input element. (#1741)